refactor: organize DTO and fix API response types
DTO organization:
- Create dto/auth/ folder for authentication DTOs
  * LoginRequest.kt
  * LoginResponse.kt
- Create dto/user/ folder for user DTOs
  * User.kt
  * SignUpRequest.kt
  * SignUpResponse.kt
  * UpdateUserRequest.kt
  * ChangePasswordRequest.kt
  * ChangePasswordResponse.kt
  * CheckUsernameResponse.kt
- Update all import statements across the project

API fixes:
- Fix AuthApiService.getCurrentUser() to return LoginResponse
- Update ProfileViewModel.getMyProfile() to handle LoginResponse
- Extract user from response correctly
