+++ app/src/main/java/com/example/active_portfolio_mobile/data/remote/api/UserPublicApiService.kt
@@ -1,7 +1,7 @@
 package com.example.active_portfolio_mobile.data.remote.api
 
-import com.example.active_portfolio_mobile.data.remote.dto.CheckUsernameResponse
-import com.example.active_portfolio_mobile.data.remote.dto.User
+import com.example.active_portfolio_mobile.data.remote.dto.user.CheckUsernameResponse
+import com.example.active_portfolio_mobile.data.remote.dto.user.User
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
