+++ app/src/main/java/com/example/active_portfolio_mobile/data/remote/dto/adventure/Adventure.kt
@@ -1,4 +1,4 @@
-package com.example.active_portfolio_mobile.data.remote.dto
+package com.example.active_portfolio_mobile.data.remote.dto.adventure
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
