+++ app/src/main/java/com/example/active_portfolio_mobile/data/remote/dto/auth/LoginRequest.kt
@@ -1,0 +1,9 @@
+package com.example.active_portfolio_mobile.data.remote.dto.auth
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LoginRequest(
+    val email: String,
+    val password: String,
+)
