+++ app/src/main/java/com/example/active_portfolio_mobile/data/remote/dto/user/ChangePasswordRequest.kt
@@ -1,0 +1,9 @@
+package com.example.active_portfolio_mobile.data.remote.dto.user
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ChangePasswordRequest(
+    val oldPassword: String,
+    val newPassword: String
+)
