+++ app/src/main/java/com/example/active_portfolio_mobile/data/remote/dto/user/ChangePasswordResponse.kt
@@ -1,0 +1,8 @@
+package com.example.active_portfolio_mobile.data.remote.dto.user
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ChangePasswordResponse(
+    val message: String
+)
