+++ app/src/main/java/com/example/active_portfolio_mobile/data/remote/dto/user/UpdateUserRequest.kt
@@ -1,0 +1,12 @@
+package com.example.active_portfolio_mobile.data.remote.dto.user
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class UpdateUserRequest(
+    val firstName: String? = null,
+    val lastName: String? = null,
+    val username: String? = null,
+    val bio: String? = null,
+    val program: String? = null,
+)
