+++ app/src/main/java/com/example/active_portfolio_mobile/ui/search/UserCard.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
-import com.example.active_portfolio_mobile.data.remote.dto.User
+import com.example.active_portfolio_mobile.data.remote.dto.user.User
 
 /**
  * Displays a user card with avatar, full name, username, and a trailing chevron icon.
