+++ app/src/main/java/com/example/active_portfolio_mobile/viewModels/AdventureSectionUpdateVM.kt
@@ -8,9 +8,9 @@ import androidx.lifecycle.ViewModel
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.viewModelScope
-import com.example.active_portfolio_mobile.data.remote.dto.AdventureSection
-import com.example.active_portfolio_mobile.data.remote.dto.AdventureSectionUpdateRequest
-import com.example.active_portfolio_mobile.data.remote.dto.Portfolio
+import com.example.active_portfolio_mobile.data.remote.dto.adventure.AdventureSection
+import com.example.active_portfolio_mobile.data.remote.dto.adventure.AdventureSectionUpdateRequest
+import com.example.active_portfolio_mobile.data.remote.dto.portfolio.Portfolio
 import com.example.active_portfolio_mobile.data.remote.api.ActivePortfolioApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
