+++ app/src/main/java/com/example/active_portfolio_mobile/viewModels/UserPortfolio.kt
@@ -2,7 +2,7 @@ package com.example.active_portfolio_mobile.viewModels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.active_portfolio_mobile.data.remote.dto.Portfolio
+import com.example.active_portfolio_mobile.data.remote.dto.portfolio.Portfolio
 import com.example.active_portfolio_mobile.data.remote.api.ActivePortfolioApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
